Fix ordering of columns whose position was changed using colReorder plugin
The colReorder plugin allows to change the position of columns in the datatable. However this changes the column index which does not match the backend index anymore. That caused that the wrong column were ordered in the backend. With this commit the original index will be put in the request, so always the correct coumn is ordered.