ExcelOpenSpoutExporter: add check for non-string values
We encountered an issue where we had a `render` callable which returned
an int. That broke the exporter because I assumed that the value is
always a string.

I had the idea (when looking at e.g. NumberColumn) that a callable given
to `render` should always return a string. But that's not very clear in
the documentation. But I also changed our `render` function to return a
string.