Fixes and changes for column filters (#369)
* Remove operator option on AbstractFilter because it is nowhere used

* AbstractFilter: get rid of $this->$key

This line is risky and inflexible as it implicitely requires each option
to have a property on the child class.

* For searchColumns, check isSearchable() in getter instead of before set

This fixes a bug which is demonstrated in the test cases included in
this commit.

This is not the most elegant solution, let me know if you have something
better.

* Make column search behaviour the same as for global search

This is the default behaviour that I would expect for a column search.