Make the raw export opt-in (#374)
The new change to skip AbstractColumn::transform() during export seems
to be a quite intrusive backwards-incompatible change, because we use
the `render` option quite a lot construct the column value (even though
we should probably use `data` for that).

This commit makes it an opt-in feature on a per-column basis.

I also add a `stripTags` option to ExcelOpenSpoutExporter.